+++ lib/stdlib/math.dart
@@ -1,0 +1,3 @@
+library dart_eval.stdlib.math;
+
+export '../src/eval/shared/stdlib/math/point.dart';